Detect unit from a Measurement value if it provided as a string
String literals have the ability to contain the unit for a
measurement. Unless a unit is explicitly defined, a string chunk between
the last space and the end is used to lookup the unit by it's abbreviation.
If the string does not match a registered abbreviation, the default unit
(magnitude) is used.